+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Add support for scalar map channel option (`-imfchan`) in 3D preview and material editor (only RGB)
 - Add support for spherical reflection map (`refl -type sphere`) in 3D preview
 - Add support for reflection map type option in material editor
+- Add filtering of material properties visibility
+- Add material preview mesh: 3x3 plane
 - Settings for MTL file editor layout separate from OBJ file editor layout
 - Setting for generating mipmaps for textures (on/off)
 - Setting for selecting face culling
@@ -49,6 +51,7 @@
 
 ### Removed
 - Remove support for platform version `221.*`
+- Remove additional light source from PBR shader
 
 ### Fixed
 - Fix error in 3D preview of a line without texture coordinates